fixing gpio issue

+++ cyclosafe/scripts/gpio.sh
@@ -2,27 +2,31 @@
 
 LOG_DIR="/home/npirard/data/logs"
 LOG_FILE="/home/npirard/data/logs/on_off.log"
-PYTHON_SCRIPT="/home/npirard/ros_ws/install/cyclosafe/share/cyclosafe/gpio.py"
+SCRIPT_DIR="/home/npirard/ros_ws/install/cyclosafe/share/cyclosafe/scripts"
+PYTHON_SCRIPT="gpio.py"
 
 mkdir -p "$LOG_DIR"
 touch "$LOG_FILE"
 
 echo "[INFO] $(date) - Powered on" | tee -a "$LOG_FILE"
+
+cd $SCRIPT_DIR
+export PYTHONPATH=$SCRIPT_DIR:$PYTHONPATH
 python3 "$PYTHON_SCRIPT"
 EXIT_CODE="$?"
 
 # Vérifier si le retour est -1 ou -2
-if [ $EXIT_CODE -eq 1 ]; then
+if [ $EXIT_CODE -eq 255 ]; then
 
 	echo "[INFO] $(date) - Shutdown triggered by button" | tee -a "$LOG_FILE"
 
-elif [ $EXIT_CODE -eq 2 ]; then
+elif [ $EXIT_CODE -eq 254 ]; then
 
    echo "[INFO] $(date) - Shutdown triggered by low battery" | tee -a "$LOG_FILE"
 
 else
 
-	echo "[WARNING] $(date) - Unknown exit code $EXIT_CODE" | tee -a "$LOG_FILE"
+	echo "[ERROR] $(date) - Unknown exit code $EXIT_CODE" | tee -a "$LOG_FILE"
 
 fi
 
@@ -31,4 +35,10 @@ sleep 5
 
 systemctl is-active --quiet service && (echo "[WARNING] $(date) - Shutdown while service still active" | tee -a "$LOG_FILE")
 
-shutdown -h now
+if [ $EXIT_CODE -eq 255 ] || [ $EXIT_CODE -eq 254 ]; then
+
+	shutdown -h now
+	# echo "SHUTDOWN"
+else
+	echo "[WARNING] $(date) - Shutdown cancelled" | tee -a "$LOG_FILE"
+fi
